Create whatsappWorkaround.js
This snippet fixes the issues with WhatsApp links. It simply looks for the UserAgent and overrides the web link for the api link on smartphones and tablets. So it should work in all desktop browsers as aspected and also should open WhatsApp on all mobile devices.
Feel free to use it.